Correctly process commit and unknown notifications
Previously, the app would halt processing notifications if there was a
commit notification; it assumed every notification was for a GH Issue or
PR.

We fix this by updating the Key for GH items to support commits as well
as issues (broadly, instead of issue number, use the commit SHA). To do
this, we move the decision on what the key should be to the relevant
Get* method and store the key directly on the GitHubItem struct.

In addition, the app will just log and skip over notifications with
unknown types -- which we determine from the api url -- instead of
halting processing of notifications.

internal/gh/github.go #